BUGFIX: Refactor to more functional and reactive paradigm
all (+2 fixed) tests passing :D

fixes test "editor doesnt crash when CodePen opener is hidden"
Previously it was relied upon, that the codepen opener component is always existant, when the codepen is open.
But it turned out, one could hide and thus destroy the component instance, and the tabValues data flow is cut.